+++ src/WebGPU/Silk.NET.WebGPU/Structs/ComputePassDescriptor.gen.cs
@@ -23,7 +23,7 @@ namespace Silk.NET.WebGPU
         (
             ChainedStruct* nextInChain = null,
             byte* label = null,
-            uint? timestampWriteCount = null,
+            nuint? timestampWriteCount = null,
             ComputePassTimestampWrite* timestampWrites = null
         ) : this()
         {
@@ -59,10 +59,10 @@ namespace Silk.NET.WebGPU
         [NativeName("Name", "label")]
         public byte* Label;
 
-        [NativeName("Type", "uint32_t")]
-        [NativeName("Type.Name", "uint32_t")]
+        [NativeName("Type", "size_t")]
+        [NativeName("Type.Name", "size_t")]
         [NativeName("Name", "timestampWriteCount")]
-        public uint TimestampWriteCount;
+        public nuint TimestampWriteCount;
 
         [NativeName("Type", "const WGPUComputePassTimestampWrite *")]
         [NativeName("Type.Name", "const WGPUComputePassTimestampWrite *")]
